Implement quality can not fall below 0

+++ src/test/java/com/gildedrose/StockTests.kt
@@ -14,8 +14,8 @@ class StockTests {
     private val initialStockList = StockList(
         lastModified = Instant.parse("2022-02-09T23:59:59Z"),
         items = listOf(
-            Item("banana", oct29.minusDays(1), 42u),
-            Item("kumquat", oct29.plusDays(1), 101u)
+            Item("banana", oct29.minusDays(1), 42),
+            Item("kumquat", oct29.plusDays(1), 101)
         )
     )
     private val fixture = Fixture(initialStockList, now = initialStockList.lastModified)
@@ -38,8 +38,8 @@ class StockTests {
         val expectedUpdatedResult = StockList(
             lastModified = now,
             items = listOf(
-                Item("banana", oct29.minusDays(1), 41u),
-                Item("kumquat", oct29.plusDays(1), 100u)
+                Item("banana", oct29.minusDays(1), 41),
+                Item("kumquat", oct29.plusDays(1), 100)
             )
         )
         assertEquals(expectedUpdatedResult, stock.stockList(now))
@@ -52,8 +52,8 @@ class StockTests {
         val expectedUpdatedResult = StockList(
             lastModified = now,
             items = listOf(
-                Item("banana", oct29.minusDays(1), 40u),
-                Item("kumquat", oct29.plusDays(1), 99u)
+                Item("banana", oct29.minusDays(1), 40),
+                Item("kumquat", oct29.plusDays(1), 99)
             )
         )
         assertEquals(expectedUpdatedResult, stock.stockList(now))
@@ -91,4 +91,4 @@ class StockTests {
     }
 }
 
-private fun updateItems(items: List<Item>, days: Int) = items.map { it.copy(quality = it.quality - days.toUInt()) }
+private fun updateItems(items: List<Item>, days: Int) = items.map { it.copy(quality = it.quality - days) }
